fix: concentration de certaines fixtures
Certaines fixtures sont réutilisées dans les tests de plusieurs packages
de "haut-niveau".
Autant les placer dans le `conftest.py` le plus haut.
